pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "veille_agent"
authors = [
  {name = "Matthieu Daviaud", email = "matthieu.daviaud@gmail.com"},
]
version = "0.1.0"
description = "Agent de veille automatique"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "beartype>=0.22.9",
    "feedparser>=6.0.12",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
veille_agent = "veille_agent.__main__:main"

[project.optional-dependencies]
dev = [
    "invoke>=2.2.0",
    "pytest-cov>=7.1.0",
    "types-pyyaml>=6.0.12",
]
docs = [
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
]
lint = [
    "mypy>=1.20.0",
    "ruff>=0.15.10",
    "types-pyyaml>=6.0.12",
]

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests", "src/veille_agent"]
addopts = [
    "--doctest-modules",
    "--cov=src/veille_agent",
    "--cov-report=term-missing",
    "--cov-report=html",
]
pythonpath = ["."]

[tool.coverage.run]
source = ["src/veille_agent"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "raise NotImplementedError",
]
skip_empty = true

[tool.mypy]
python_version = "3.13"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src/veille_agent"]

[tool.hatch.build.targets.wheel.force-include]
"src/veille_agent/py.typed" = "veille_agent/py.typed"
"src/veille_agent/config/profile.yaml" = "veille_agent/config/profile.yaml"

[dependency-groups]
lint = [
    "mypy>=1.20.0",
    "ruff>=0.15.10",
]
